requirements.txt
streamlit
pandas
python-dotenv
watchdog
plotly
databricks-sql-connector